Setup github workflows for CI.
The following five workflows are set up to build and install O2SCL as a
shared library, and then run make check:
* openSUSE Tumbleweed i386: Fast suite of tests (make check fails).
* openSUSE Tumbleweed x86_64: Fast suite of tests (successful).
* openSUSE Tumbleweed x86_64: Full suite of tests (make check fails).
* Ubuntu x86_64: Fast suite of tests (successful).
* Ubuntu x86_64: Full suite of tests (successful).

All tests are confgured to build with python bindings disabled as this
cannot be currently tested in this set-up anyway. Where possible ccache
is used to speed up repeated compilation of time-taking make and make
check.